Second task works but incorrectly
It should output data without second line(text_with_trash) and it also should exlude common english words such as prepositions and so on. Now it only shows data without trash signs but with common words ('was', 'let', 'is') But i don't know how to solve this without using DataFrame and StopWordsRemover. And also the first line "When Israel was in Egypt's land" disappeared.